Move DB, Meas, and Field List components to /shared
